made it more like a website than a one page rager
