+++ pages/manufacturing.html
@@ -5,14 +5,14 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <meta name="description" content="">
   <title>Dothard & Simbleton</title>
-  <link rel="stylesheet" type="text/css" href="./styles/main.css">
+  <link rel="stylesheet" type="text/css" href="../styles/main.css">
 </head>
 <body>
   <header>
     <ul>
-      <li><a href="#">Active Businesses</a></li>
-      <li><a href="./pages/newYork.html">New York Businesses</a></li>
-      <li><a href="./pages/manufacturing.html">Manufacturing Businesses</a></li>
+      <li><a href="../index.html">Active Businesses</a></li>
+      <li><a href="./newYork.html">New York Businesses</a></li>
+      <li><a href="#">Manufacturing Businesses</a></li>
     </ul>
   </header>
   <main id="main">
@@ -23,6 +23,6 @@
   <footer>
   </footer>
 
- <script type="module" src="./scripts/main.js"></script>
+ <script type="module" src="../scripts/manufacturers.js"></script>
 </body>
 </html>
