+++ pages/newYork.html
@@ -1,0 +1,27 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <meta name="description" content="">
+  <title>Dothard & Simbleton</title>
+  <link rel="stylesheet" type="text/css" href="../styles/main.css">
+</head>
+<body>
+  <header>
+    <ul>
+      <li><a href="../index.html">Active Businesses</a></li>
+      <li><a href="#">New York Businesses</a></li>
+      <li><a href="./manufacturing.html">Manufacturing Businesses</a></li>
+    </ul>
+  </header>
+  <main id="main">
+    <div class="businessList--newYork" id="businesses--newYork"></div>
+    
+  </main>
+  <footer>
+  </footer>
+
+ <script type="module" src="../scripts/NewYork.js"></script>
+</body>
+</html>
