+++ scripts/NewYork.js
@@ -1,0 +1,7 @@
+import { NewYorkBusinessList } from "./NewYorkBusinessList.js"
+
+
+const newYorkBusinessesContainer = document.getElementById("businesses--newYork")
+
+
+newYorkBusinessesContainer.innerHTML = NewYorkBusinessList()
