+++ scripts/manufacturers.js
@@ -1,0 +1,8 @@
+import { ManufacturingBusinessList } from "./ManufacturingBusinessList.js"
+
+
+const manufacturingBusinessContainer = document.getElementById("businesses--manufacturing")
+
+
+
+manufacturingBusinessContainer.innerHTML = ManufacturingBusinessList()
