+++ styles/main.css
@@ -16,6 +16,13 @@ h1,h2,h3,h4,h5,h6 {
   letter-spacing: 2px;
 }
 
+/* Main Styles
+-------------------------------------------- */
+/* #main {
+  display: flex;
+  justify-content: space-around;
+} */
+
 /* Business Styles 
 ---------------------------------------------*/
 .business {
